Refactor yolo task widget
